fix: update patching mode (#2849)
* feat: test update models

* feat: map for execution update

* feat: mapping for executor and test sources

* feat: mapping for test suite update

* fix: support custom secret update

* fix: change method call parameters

* fix: chnage executor update client method

* feat: test cli update command

* feat: test suite cli update

* fix: change make cmd

* fix: dep update

* fix: return webhook object

* fix: change comment

* fix: condition typo

* fix: change flag typos

* fix: change method call params

* fix: go fmt

* fix: type name typo

* fix: change pointer ops

* fix: typo in param name

* fix: dep update

* fix: check for empty struct

* fix: send empty content fields

* fix: help typo

* fix: check for empty repo and execution

* fix: dep update

+++ pkg/mapper/testsuites/kube_openapi.go
@@ -58,7 +58,7 @@ func mapCRStepToAPI(crstep testsuitesv2.TestSuiteStepSpec) (teststep testkube.Te
 	case crstep.Delay != nil:
 		teststep = testkube.TestSuiteStep{
 			Delay: &testkube.TestSuiteStepDelay{
-				Duration: int32(crstep.Delay.Duration),
+				Duration: crstep.Delay.Duration,
 			},
 		}
 	}
